% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{edge_summary}
\alias{edge_summary}
\title{Summarize the edge information from fitted result}
\usage{
edge_summary(x, ...)
}
\description{
This function calculates edge strength from the MCMC sampling results.
}
