% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{edge_summary.structure}
\alias{edge_summary.structure}
\title{Summarize the edge information from fitted result}
\usage{
\method{edge_summary}{structure}(result_structure, exposure = NULL)
}
\description{
This function calculates edge strength from the MCMC sampling results.
}
