% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data_for_outcome}
\alias{example_data_for_outcome}
\title{Example data for outcome learning}
\format{
A list with 4 datasets
}
\source{
DepMap
}
\usage{
example_data_for_outcome
}
\description{
Multi-omics (CNA, mRNA, Protein) data of lung cancer cell lines within the RTK pathway, along with Erlotinib drug response (Drug) data.
}
\keyword{datasets}
