% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data_for_structure}
\alias{example_data_for_structure}
\title{Example data for structure learning}
\format{
A list with 3 datasets
}
\source{
DepMap
}
\usage{
example_data_for_structure
}
\description{
Multi-omics (CNA, mRNA, Protein) data of lung cancer cell lines within the RTK pathway.
}
\keyword{datasets}
