% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_result_outcome}
\alias{example_result_outcome}
\title{Example fitted result from outcome model learning}
\format{
A list containing fitting information
}
\usage{
example_result_outcome
}
\description{
Output of GFusionMed::fit_outcome_model(data_for_outcome)).
}
\keyword{datasets}
