% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_result_structure}
\alias{example_result_structure}
\title{Example fitted result from structure learning}
\format{
A list containing fitting information
}
\usage{
example_result_structure
}
\description{
Output of GFusionMed::fit_structure_model(example_data_for_structure).
}
\keyword{datasets}
