% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{fit_outcome_model}
\alias{fit_outcome_model}
\title{Outcome Model Learning}
\usage{
fit_outcome_model(
  y,
  X,
  gamma.prob = 0.1,
  lambda,
  delta,
  burnin.S,
  inf.S,
  model = "normal"
)
}
\description{
This function prints the nudd.
}
