% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{fit_structure_model}
\alias{fit_structure_model}
\title{Structure Learning}
\usage{
fit_structure_model(
  v.ch,
  v.pa,
  Y,
  eta.prob = 0.1,
  gamma.prob = 0.1,
  lambda,
  delta,
  burnin.S,
  inf.S
)
}
\description{
This function prints the nu.
}
