% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{fit_structure_model_temp}
\alias{fit_structure_model_temp}
\title{Internal function}
\usage{
fit_structure_model_temp(
  v.ch,
  v.pa,
  Y,
  eta.prob = 0.3,
  gamma.prob = 0.3,
  lambda,
  delta,
  burnin.S,
  inf.S
)
}
\description{
This is an internal function that is not exported.
}
