% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFusionMed-package.R
\docType{package}
\name{GFusionMed-package}
\alias{GFusionMed-package}
\alias{GFusionMed}
\title{GFusionMed: A Tool for Bayesian Graphical Data Fusion and Mediation Analysis in Pharmaco-omics}
\description{
GFusionMed is an R package designed for Bayesian graphical data fusion and mediation analysis, 
enabling the integration of multi-omics data. By leveraging multi-layered Gaussian graphical models 
and advanced mediation analysis methods, GFusionMed helps researchers explore complex biological 
relationships and uncover mediating mechanisms underlying pharmacological responses.
}
\details{
The package offers an efficient pipeline for structure learning and mediation analysis, 
promoting accessibility to sophisticated statistical methods for pharmaco-omics research.

For a complete list of functions, use:
\code{library(help = "GFusionMed")}
}
\author{
GFusionMed authors
}
\keyword{package}
