% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{perform_mediation_analysis}
\alias{perform_mediation_analysis}
\title{Perform Mediation Analysis}
\usage{
perform_mediation_analysis(
  palist,
  chlist,
  fit.med,
  fit.out,
  ci = c(0.025, 0.975)
)
}
\description{
This function prints.
}
