% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_network.structure_outcome}
\alias{plot_network.structure_outcome}
\title{Internal function}
\usage{
plot_network.structure_outcome(
  result_structure,
  result_outcome,
  exposure,
  path,
  file_name,
  width = 20,
  height = 11,
  vector_layer_color = NULL,
  seed = 1234
)
}
\description{
This is an internal function that is not exported.
}
