% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Sankey Plot}
\usage{
plot_sankey(
  edge_structure,
  edge_outcome,
  group_information,
  vector_layer_color = NULL
)
}
\description{
Visualizes the network when group information for the variables is available.
}
