% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{print.exposure_df}
\alias{print.exposure_df}
\title{Internal function}
\usage{
\method{print}{exposure_df}(x, ..., digits = 6, n_head = 10, show_all = FALSE)
}
\description{
This is an internal function.
}
