% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateDirected}
\alias{updateDirected}
\title{Internal function}
\usage{
updateDirected(
  y,
  X,
  lambda,
  delta,
  CCinv,
  B,
  Gamma,
  kappa,
  qmat,
  v.no.ue,
  v.aa,
  no.tau
)
}
\description{
This is an internal function that is not exported.
}
