% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateUndirected}
\alias{updateUndirected}
\title{Internal function}
\usage{
updateUndirected(
  v,
  dat,
  lambda,
  delta,
  Alpha,
  eta,
  kappa,
  pmat,
  no.de,
  bCb,
  no.tau
)
}
\description{
This is an internal function that is not exported.
}
