% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{v.updateUndirected}
\alias{v.updateUndirected}
\title{Internal function}
\usage{
v.updateUndirected(
  y,
  X,
  lambda,
  delta,
  Alpha,
  eta,
  kappa,
  pmat,
  no.de,
  bCb,
  no.tau
)
}
\description{
This is an internal function that is not exported.
}
