% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{fit_outcome_model}
\alias{fit_outcome_model}
\title{Outcome model learning}
\usage{
fit_outcome_model(
  input_list,
  lambda = 5,
  delta = 2,
  burnin.S = 10000,
  inf.S = 10000,
  gamma.prob = 0.3,
  seed = 1234,
  model = "normal"
)
}
\description{
Fits an outcome model to study the relationships between omics data and a specific outcome (e.g., drug response), compatible with various types of outcomes (continuous, binary, ordinal). Set model = "normal" if the data is continuous, and set model = "probit" otherwise.
}
