% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{fit_structure_model}
\alias{fit_structure_model}
\title{Multi-omics structure learning}
\usage{
fit_structure_model(
  input_list,
  lambda = 5,
  delta = 2,
  burnin.S = 10000,
  inf.S = 10000,
  eta.prob = 0.3,
  gamma.prob = 0.3,
  seed = 1234,
  cores = parallel::detectCores() - 1,
  parallel = FALSE,
  method = "AND"
)
}
\description{
Constructs a multi-layered graphical model to capture both undirected (within-layer) and directed (between-layer) relationships in multi-omics data. By default, multiple CPU cores are used. It is recommended to allocate as many cores as the number of layers in the input.
}
