% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{perform_mediation_analysis}
\alias{perform_mediation_analysis}
\title{Mediation Analysis}
\usage{
perform_mediation_analysis(result_structure, result_outcome, exposure = NULL)
}
\description{
Conducts mediation analysis to identify indirect effects of omics variables on outcomes, revealing potential causal pathways between multi-omics data and the observed pharmacological response.
}
