% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_network}
\alias{plot_network}
\title{Network Visualization}
\usage{
plot_network(
  result_structure = NULL,
  result_outcome = NULL,
  exposure = NULL,
  path,
  file_name,
  width = 20,
  height = 11,
  vector_layer_color = NULL,
  seed = 1234
)
}
\description{
Visualizes the network, where edge transparency reflects the strength of inferred relationships.
}
